backend/requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==7.1.0
djangorestframework==3.15.1
phonenumbers==9.0.5
requests==2.31.0
geopy==2.4.1
rollbar==1.3.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
gunicorn==23.0.0
